Move unit test under tests/. Now runs.
